[1] Added a widget to the main window to list loaded files.
The idea here is to have a list of the files that are loaded in the application.
For now, it lists all files that are dropped into the main window - later,
it will only show executable files.

I use drag-and-drop to bring in files into the application because it is
convenient (more so than using a file dialog). It is the main window that
handles the drag-and-drop events; ideally, it would be the list widget, but
that would require either using an event filter or subclassing the list widget
to handle the drag-and-drop events, both of which require more work than
I am willing to put in at this point.

The top-level window is now also implemented as a `QMainWindow` sub-class
because I feel that I will eventually require more top-level control.
This, however, requires setting `CMAKE_AUTOMOC` to `ON` in the `CMakeLists.txt.

The main visible top-level container is a `QSplitter` since at this point
in development it will probably be convenient to be able to widen the list
widget.