[2] Create a skeleton widget to show contents of a single executable file
The `EXEViewer` class has been added to show the contents of an executable
file (specifically, a .exe file). This class is a subclass of `QTabWidget`
because it will eventually be fleshed out to show different contents of
a .exe in separate tabs.

The viewer side of the application has also been set up, as a stacked widget
that would show the viewer widgets based on what is selected in the loaded
files list. For now, the only available viewer is `EXEViewer`, but that
will change in the future.

The stretch factor for the stacked widget has been set on the top-level
splitter so that the stacked widget will take up the majority of the available
space.