[4] Set up groundwork for loading a .exe file
Some code has been added to eventually be able to load a .exe file.
For this, the `EXEFile` structure has been created to hold the contents
of the file, and this is passed to `EXEViewer` to display the contents.
The function `loadEXEFile/1` has been added (but left as a skeleton) to
actually load a .exe file (it will be fleshed out later).

`EWEAMainWindow::dropEvent/1` has been modified to have the following events
in the correct sequence:
- Load the contents of the .exe file
- Display the contents of the .exe file
- Add the path of the .exe file to the loaded files list
The important detail here is that we don't want create/modify widgets if
we cannot load the file.

`EXEViewer::setUpRawDataTab/0` has also been set up to display some dummy
data. This function, too, will be fleshed out later.

The C++ standard for the application has been set to C++20. This was intially
done to enable the use of ranged-for loops with initializer, but now it's
there just because.