[6] Show headers of a .exe file (1)
This commits adds just enough code to carve out the .exe headers from the
file contents (which were loaded in the previous commit) and display them
in a new "Headers" tab. For now, only PE32+ files are supported.

Most of the code goes into setting up the widgets to display header
information. For now, only one field is shown from each header; more will
be added as needed (also, the current choice of shown fields is arbitrary).
The same applies to the structs used to store the header information - for
now, they only contain blobs of bytes (mostly), but those blobs will be
divided into dedicated fields as needed.