[7] Show headers of a .exe file (2)
Added widgets to show more fields in the .exe headers. The choice of the
fields is, again, arbitrary, since the fields are not really being used
for anything yet.

The data directory entries in the NT optional header are now also shown,
with the ones corresponding to the (likely) missing directories being
disabled.

It is worth nothing that the sort of code added in this commit is quite
tedious to write, even with Copilot.

The "Headers" tab has also been renamed to "File Headers" to distinguish
the name from "section headers".

ASIDE: The PE format documentation from Microsoft is not very clear on
       whether or not the number of data directory entries can be less
       than 16, or on what to do when that happens (as far as I can tell,
       the standard indices of the directory entries are the only way to
       distinguish between them).