[13] Extract and show export table
What the title says. This change is mainly intended for DLLs, since .exe
files tend to not export functions.

This commit is a first step: it only extracts the export name table and
lists all of the functions in it in the "Exports" tab. Extracting the
export address table and the ordinal table is for subsequent commits,
and it will involve some extra effort since as far as I know, the three
tables in question could be in different sequences.

Note that the export table is supposed to be present in the .edata section -
however, as with the import table, we can probably not assume that the
section hosting the export table is named ".edata".