[18] Refactor some PE format handling code (3)
Renamed PEFile.h/.cpp to PEFormat.h/.cpp and EXEFile.h/.cpp to PEFiles.h/.cpp.
The goal to have have PEFormat.h/.cpp to handle the actual PE format, and
PEFiles.h/.cpp to use the stuff in PEFormat.h/.cpp to actually read in files
that are in the PE format.