[19] Enabled loading .obj files and showing some basic info
Made it possible to load .obj files and show the NT file header and the
section headers. The implementation is in line with that for .exe and .dll
files, one big difference being that .obj files have multiple sections with
identical names (from what I understand, these are combined into a single
section with the same name, though I have no idea what the sequence should be).